fixes and improvements in preparation for testing

+++ products/src/controllers/products.controller.ts
@@ -1,4 +1,5 @@
 import { Request, Response } from "express";
+import mongoose from "mongoose";
 
 import Product from "../models/product";
 
@@ -10,14 +11,17 @@ class ProductController {
   }
 
   async getProduct(req: Request, res: Response) {
-    const product = await Product.findById(req.params.id);
+    if (mongoose.Types.ObjectId.isValid(req.params.id)) {
+      const product = await Product.findById(req.params.id);
 
-    if (!product) {
-      return res
-        .status(400)
-        .json(`Product with id: ${req.params.id} not found!`);
+      if (!product) {
+        return res
+          .status(400)
+          .json(`Product with id: ${req.params.id} not found!`);
+      }
+      return res.json(product);
     }
-    return res.json(product);
+    return res.status(400).json("Invalid ID");
   }
 
   async createProduct(req: Request, res: Response) {
@@ -27,15 +31,21 @@ class ProductController {
   }
 
   async updateProduct(req: Request, res: Response) {
-    const result = await Product.findByIdAndUpdate(req.params.id, req.body);
+    if (mongoose.Types.ObjectId.isValid(req.params.id)) {
+      const result = await Product.findByIdAndUpdate(req.params.id, req.body);
 
-    res.json(result);
+      res.json(result);
+    }
+    return res.status(400).json("Invalid ID");
   }
 
   async deleteProduct(req: Request, res: Response) {
-    const result = await Product.findByIdAndDelete(req.params.id);
+    if (mongoose.Types.ObjectId.isValid(req.params.id)) {
+      const result = await Product.findByIdAndDelete(req.params.id);
 
-    res.json(result);
+      res.json(result);
+    }
+    return res.status(400).json("Invalid ID");
   }
 }
